feat(mealplan): month transition handling for continuous meal plans (#381)
This feature ensures users always have continuous meal plans without manual
month switching. When generating meal plans at month-end, the system now
automatically extends into the next month to maintain a minimum of 4 future
weeks.

The generation logic detects when fewer than 4 weeks remain in the current
month and seamlessly continues into the next month. Users see a clear
transition message like "Generating weeks for rest of October + first 2
weeks of November" so they understand their meal plans span multiple months.

Week metadata now includes month and year information for accurate calendar
display and filtering. This enables the UI to properly organize and present
meal plans across month boundaries.

+++ crates/mealplan/tests/generate.rs
@@ -83,9 +83,9 @@ async fn test_random() -> anyhow::Result<()> {
         .await?;
 
     let now = OffsetDateTime::now_utc();
-    let weeks = imkitchen_mealplan::next_four_mondays(now.unix_timestamp())?;
+    let weeks = imkitchen_mealplan::next_four_mondays(now.unix_timestamp() as u64)?;
     for week in weeks {
-        let row = query.find(week as u64, "john").await?.unwrap();
+        let row = query.find(week, "john").await?.unwrap();
         assert!(row.slots.is_empty());
         assert_eq!(row.status.0, Status::Processing);
     }
@@ -101,7 +101,7 @@ async fn test_random() -> anyhow::Result<()> {
         .await?;
 
     for week in weeks {
-        let row = query.find(week as u64, "john").await?.unwrap();
+        let row = query.find(week, "john").await?.unwrap();
         assert!(!row.slots.is_empty());
         assert_eq!(row.status.0, Status::Idle);
     }
